Added login ability
Added a few login methods wrote the actual login action and added support for BCrypt login and storing into the databse as well as the database file dump so you can import it to your own databse. I also added tamper detection to sniff out and close Packet capturing applications to stop those pesky hackers :) that being said this is not a working version yet.