Added a good bit this update :)
I added an admin panel where you will generate a auth token for your program as a security messure to keep people from accessing your database with the dll. I added a program statistics class that allows you to get the number of users you have online. I also added in a logout class to update users online status again as well as new additions to the database on such as registered date and time and lastlogin date and time. A video detailing mhow to set thi sall up will be happening soon!