Added a License capabilities (Update: 2.0.1.0)
In this update I made two seperate registration methods a free one which is the registration i have already demonstrated to you guys and then I added a paid registration method that allows you to only register if they have a license that will be generated via the Admin Panel. This also means the Database file was updated to support these features as well as making it possible to see all licenses, redeemed licenses and who they were redeemed by as well as unredeemed licenses in a list view within the admin panel as well as change the redeemed status and delete licenses from the db! Hopefully you guys enjoy this update haven't had a lot of time to work on NETLock but i grinded this out tonight so feel free to leave feedback!