Version 3.0
In this update I made error messages able to be appended as text to messagebox's and other text related viewers through the property info accessable in LoginResponse, LogoutReponse, and RegisterReponse. Examples have been added to all example applications. Also I added the ability to copy licenses from the License ListView. The file add on in this update is a MD5 Hash calculator that will aid you in using InnerLock.cs to protect your applications integrity by compairing the "CurrentMD5" variable to a live calculated MD5 hash of NETLock.dll!